Added getters for AppRequest object and fixed a bug with the id field

+++ requests-microservice/src/main/java/nl/tudelft/sem/template/requests/domain/RequestRepository.java
@@ -12,5 +12,5 @@ public interface RequestRepository extends JpaRepository<AppRequest, String> {
     /**
      * Find request by id.
      */
-    Optional<AppRequest> findById(int id);
+    Optional<AppRequest> findById(long id);
 }
